Update ModelTrainer to have the list of losses computed in training

+++ RealTests/TimeSeriesForecasting.NeuralNetwork/IModelTrainer.cs
@@ -5,7 +5,7 @@ namespace TimeSeriesForecasting.NeuralNetwork
     public interface IModelTrainer
     {
         public bool IsTrained { get; }
-        public double CurrentLoss { get; }
+        public float CurrentLoss { get; }
         public void Fit(Tensor x, Tensor y, int epochs);
     }
 }
